% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grade.R
\name{calculate_grade}
\alias{calculate_grade}
\title{When all course components are presented,
calculate the course overall grade based on information provided.}
\usage{
calculate_grade(input_file_path)
}
\arguments{
\item{input_file_path}{(string) :
Path to read the saved .csv file as a string.
Under the current working directory to avoid permission issue.
The path should include the file name.}
}
\value{
return_msg (string) :
The course overall grade once all course components present
when all grades are present.
Otherwise return warning message.
}
\description{
When all course components are presented,
calculate the course overall grade based on information provided.
}
\examples{
None
}
