% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_course.R
\name{construct_course}
\alias{construct_course}
\title{Allow user to input course components' information
(component name and coresponding weight) one by one.
Course components could be the following for example:
1. Assignment(s)
2. Quiz(s)
3. Lab(s)
4. Midterm(s)
5. Final(s)
6. Any user self-defined course component name(s)}
\usage{
construct_course(course_name, output_file_path)
}
\arguments{
\item{course_name}{the course name as a string to be used as the output file name}

\item{output_file_path}{path to the output .csv file to be saved}
}
\value{
None
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Save the course information as a .csv file
(named based on the 'course_name' paramater)
to the file path specified by 'output_file_path' paramater
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
construct_course('dsci524','/') 
}
}
