% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradecalculatorr-package.R
\docType{package}
\name{gradecalculatorr-package}
\alias{gradecalculatorr}
\alias{gradecalculatorr-package}
\title{gradecalculatorr: Calculates Grades For a Course}
\description{
This python package calculates grades for a course and allows users to customize their own course information with self-defined course component names, to update grades for different course components, and even understand how well the final exam needs to be to pass the course or achieve a target final grade.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yukunzGIT/gradecalculatorr}
  \item Report bugs at \url{https://github.com/yukunzGIT/gradecalculatorr/issues}
}

}
\author{
\strong{Maintainer}: Shirley Zhang \email{wenlansz@student.ubc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Edward Yukun Zhang \email{ykzhang1211@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Chen Lin \email{chen.lin.0404@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
