% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_grades.r
\name{load_course}
\alias{load_course}
\title{Loads the 'course.csv' file}
\usage{
load_course(input_file)
}
\arguments{
\item{input_file}{path to the 'course.csv' file}
}
\value{
data.frame which holds the information from 'course.csv'
}
\description{
Takes the path to the course's .csv file as an argument ('course.csv'), reads the file,
and saves the file as a dataframe to be returned
}
\examples{
\dontrun{
load_course('/tests/testthat/dummycourse.csv')
}
}
