% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_final.R
\name{predict_final}
\alias{predict_final}
\title{Calculate what grade the student needs to have for the final assignment in order to reach the goal for overall grade
and return the scores needed for the final assignment.}
\usage{
predict_final(input_file, goal)
}
\arguments{
\item{input_file}{A string path to read the saved .csv file as a string.}

\item{goal}{A numeric value to set the goal for your overall grade for this course.}
}
\value{
final_score_needed which is the scores needed for the final assignment. Otherwise return warning message.
}
\description{
Calculate what grade the student needs to have for the final assignment in order to reach the goal for overall grade
and return the scores needed for the final assignment.
}
\examples{
\dontrun{
predict_final('/tests/testthat/dsci524_grades_all.csv', 80.0)
}
}
