% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_grades.r
\name{save_course_csv}
\alias{save_course_csv}
\title{Saves the updated course information to a .csv file}
\usage{
save_course_csv(updated_course_df, output_file)
}
\arguments{
\item{updated_course_df}{updated dataframe for the course to be saved as a .csv file}

\item{output_file}{path to the output .csv file}
}
\value{
None
}
\description{
Takes the data.frame of the course with updated grades, the path to the output
file, and saves the data.frame as a .csv
}
\examples{
\dontrun{
save_course_csv(df, 'Assignment 3', '/tests/testthat/dummycourse.csv')
}
}
