% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_grades.r
\name{update_component_grade}
\alias{update_component_grade}
\title{Saves the updated course information to a .csv file}
\usage{
update_component_grade(input_file, component, grade)
}
\arguments{
\item{input_file}{path to the 'course.csv' file}

\item{component}{name of the component (or assignment) of the course to be updated}

\item{grade}{numeric grade of the component (or assignment) to be inserted into the data.frame}
}
\value{
None
}
\description{
Takes the data.frame of the course with updated grades, the path to the output
file, and saves the data.frame as a .csv
}
\examples{
\dontrun{
update_component_grade('/tests/testthat/dummycourse.csv', 'Assignment 3', 97)
}
}
