% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grade.R
\name{calculate_grade}
\alias{calculate_grade}
\title{When all course components are presented,
calculate the course overall grade based on information provided.}
\usage{
calculate_grade(input_file_path)
}
\arguments{
\item{input_file_path}{Path to read the saved .csv file as a string.}
}
\value{
return_msg The course overall grade once all course components present
}
\description{
When all course components are presented,
calculate the course overall grade based on information provided.
}
\examples{
NA
}
